[coro_rpc] fix header name
