+++ include/coro_rpc/coro_rpc_server.hpp
@@ -15,4 +15,4 @@
  */
 #pragma once
 #include <coro_rpc/coro_rpc/coro_rpc_server.hpp>
-#include <coro_rpc/coro_rpc/router_impl.hpp>
+
